* mmintrin.h (__m64): typedef it to v2si. 	(_mm_cvtsi32_si64, _mm_cvtsi32_si64_mm_sll_pi16, 	_mm_sll_pi32, _mm_sll_pi64, _mm_slli_pi64, _mm_sra_pi16, 	_mm_sra_pi32, _mm_srl_pi16, _mm_srl_pi32, _mm_srl_pi64, 	_mm_srli_pi64, _mm_and_si64, _mm_andnot_si64, 	_mm_or_si64, _mm_xor_si64): Add neccesary casts. 	* xmmintrin.h (_mm_setzero_si64): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@58316 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/i386/xmmintrin.h
@@ -1037,7 +1037,7 @@ _mm_prefetch (void *__P, enum _mm_hint __I)
 static __inline void
 _mm_stream_pi (__m64 *__P, __m64 __A)
 {
-  __builtin_ia32_movntq (__P, __A);
+  __builtin_ia32_movntq (__P, (long long)__A);
 }
 
 /* Likewise.  The address must be 16-byte aligned.  */
